refactor: data type 및 tree 관련 함수 이전
